Consolidate user service tests
